Move geotiffwriter to own package; extend vector docs for extent
