+++ raster/src/main/scala/geotrellis/raster/io/geotiff/writer/GeoTiffWriter.scala
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package geotrellis.raster.io.geotiff
+package geotrellis.raster.io.geotiff.writer
 
 import geotrellis.raster._
 import geotrellis.raster.io._
+import geotrellis.raster.io.geotiff._
 import geotrellis.vector.Extent
 import geotrellis.proj4.CRS
