+++ vector-test/src/test/scala/spec/geotrellis/vector/op/affine/GeometryCollectionTransformationMethodsSpec.scala
@@ -1,4 +1,4 @@
-package geotrellis.vector.affine
+package geotrellis.vector.op.affine
 
 import geotrellis.vector._
 import geotrellis.testkit.vector._
