+++ vector/src/main/scala/geotrellis/vector/README.md
@@ -9,6 +9,7 @@ In addition to working with raster data, Geotrellis provides a number of facilit
 `Feature`s correspond to the objects listed  under `features` in a geojson `FeatureCollection`. `Geometry`s, to `geometries` in a geojson `Feature`.
 
 ##Geometries
+
 The base `Geometry` class can be found in `Geometry.scala`. Concrete geometries include:
 + `geotrellis.vector.Point`
 + `geotrellis.vector.MultiPoint`
@@ -56,11 +57,11 @@ myML.unionGeometries
 ```
 
 The following packages extend `Geometry` capabilities:
-- [geotrellis.vector.affine](affine/)
 - [geotrellis.vector.io.json](io/json/)
 - [geotrellis.vector.io.WKT](io/WKT/)
 - [geotrellis.vector.io.WKB](io/WKB/)
 - [geotrellis.vector.op](op/)
+- [geotrellis.vector.op.affine](op/affine/)
 - [geotrellis.vector.reproject](reproject/)
 
 ##Features
@@ -106,9 +107,13 @@ def readFeatureJson[D: JsonReader, G <: Geometry: JsonReader, F <: Feature[D]](v
 Right away, `D`'s purpose is clear: any `D` which comes with the tools necessary for json serialization and deserialization will suffice. In effect, `data` corresponds to the "properties" member of the geojson spec's Feature object.
 If you can provide the serialization tools (almost certainly implicit conversions between some case class and [spray json](https://github.com/spray/spray-json)), the `Feature` object in `geotrellis.vector` does the heavy lifting of embedding your (thus serializable) data into the larger structure which includes a geometry. There's even support for geojson IDs: the "ID" member of a geojson Feature is represented by the keys of a `Map` from `String` to `Feature[D]`. Data in both the ID and non-ID variants of geojson Feature formats is easily transformed.
 
-###Submodules
-These submodules define useful methods for dealing with the denizens of `geotrellis.vector`:
+##GeoTrellis Extents
+
+There's one more piece to the `geotrellis.vector` puzzle: `Extent`. a `geotrellis.vector.Extent` is nothing more than a rectangular polygon which is projected (look [here](../../../../../proj4/src/main/scalageotrellis/proj4) for more on projection). This is useful mainly as a tool for defining the extent covered by a tile - the two jointly yield a raster (for more on rasters, go [here](../../../../../raster/src/main/scalageotrellis/raster)).
+
+##Submodules
+
+These submodules define useful methods for dealing with the entities that call `geotrellis.vector` home:
 `geotrellis.vector.io` defines input/output (serialization) of geometries
 `geotrellis.vector.op` defines common operations on geometries
 `geotrellis.vector.reproject` defines methods for translating between projections
-`geotrellis.vector.affine` defines transformations which preserve collinearity
